feat: sistema de sanciones por dias de demora + 3 strikes = 1 mes de suspension

+++ src/com/bibliotech/exception/PrestamoNoEncontradoException.java
@@ -1,7 +1,7 @@
 package com.bibliotech.exception;
 
 public class PrestamoNoEncontradoException extends BibliotecaException {
-    public PrestamoNoEncontradoException(int id) {
+    public PrestamoNoEncontradoException(long id) {
         super("No se encontró un préstamo con ID: " + id);
     }
 }
